Move the message bus into service/
 - get rid of the standalone message bus
 - get rid of the standalone "instance DB"
