+++ remote/metrics.go
@@ -4,14 +4,12 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/go-kit/kit/metrics"
 	"github.com/go-kit/kit/metrics/prometheus"
 	stdprometheus "github.com/prometheus/client_golang/prometheus"
 
 	"github.com/weaveworks/flux"
 	"github.com/weaveworks/flux/job"
 	fluxmetrics "github.com/weaveworks/flux/metrics"
-	"github.com/weaveworks/flux/service"
 	"github.com/weaveworks/flux/update"
 )
 
@@ -131,24 +129,4 @@ func (i *instrumentedPlatform) GitRepoConfig(regenerate bool) (_ flux.GitConfig,
 		).Observe(time.Since(begin).Seconds())
 	}(time.Now())
 	return i.p.GitRepoConfig(regenerate)
-}
-
-// BusMetrics has metrics for messages buses.
-type BusMetrics struct {
-	KickCount metrics.Counter
-}
-
-var (
-	BusMetricsImpl = BusMetrics{
-		KickCount: prometheus.NewCounterFrom(stdprometheus.CounterOpts{
-			Namespace: "flux",
-			Subsystem: "bus",
-			Name:      "kick_total",
-			Help:      "Count of bus subscriptions kicked off by a newer subscription.",
-		}, []string{}),
-	}
-)
-
-func (m BusMetrics) IncrKicks(inst service.InstanceID) {
-	m.KickCount.Add(1)
 }
